server-python/pyproject.toml
[project]
name = "ytllm-server"
version = "1.0.0"
description = "YouTube AI Chat Agent - Python Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "google-generativeai>=0.8.0",
    "chromadb>=0.5.0",
    "fpdf2>=2.7.0",
    "python-docx>=1.1.0",
    "youtube-transcript>=1.1.0",
    "slowapi>=0.1.9",
    "sentry-sdk>=2.0.0",
    "fastapi-cors>=1.0.0",
    "cachetools>=5.5.0",
    "python-dotenv>=1.0.0",
    "xmltodict>=0.13.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = []
